Use branch name shorthand
Previous branch name required manual parsing to find the name at
the end of the full reference by splitting on the backslash;
however, this breaks if there are / in the name of the branch,
which is a practice used by some developers to group branchnames.

git2 has the `shorthand` API call to solve this issue, which
removes the need to manually parse the name, making branch name
and upstream name robust to backslashes in the name of the
repository.