Apply a patch to rust-ansi-term to fix line wrap issue
Manually add escapes to code which prints color sequences; the
problem is that these sequences are only properly interpreted in
the shell prompt and are printed as the weird block-y unicode
things when they appear in _any_ other context.  As a result, this
is only a temporary fix while I look into better methods of adding
the escapes conditionally, perhaps using command-line options.

Fixes #10.