Remove backslashes from most parts of language
Backslashes were added to accomodate multi-character named args,
but since I have no intention of adding them anymore, they are now
redundant and just make glitter harder to use.

Additionally, remove the `\\` command which would print a literal
backslash, as '\' does the same thing, meaning the only required
escape is the `\'` command.  All other characters can be added with
literals like '#' for example.

The only expression that still requires a backslash is the \() group
to avoid ambiguity with the other contexts of (), but this means that
the other group expressions [],<>,{}, now can be just written as is
with matching automatically functioning for them without annoying
backslashes.