Upgrade parser and simplify language
Remove unecessary vectors from the AST, opting to use the newly
structured CompleteStyle (formerly StyleContext) rather than a
vector of styles.  This is lossy, but more efficient.

Now, styles are written like so:

    #g_*('test')

As opposed to the old form:

    #g;_;*('test')

Further, in the upgrade to nom 5, switch from macro-based to
function/closure based implementation.  Work is still necessary to
properly propagate errors out of the parser and display them to
the user in a friendly way.