Add separator support
Separators can appear between expressions and will intelligently
apper or not appear depending on whether the previous and follwing
separators are drawn.  For example:

    "[#g(MARD)|#r(mard)]"

The "|" character is a separator and will only appear if an item
from the "MARD" group and the "mard" group _both_ appear, then the
"|" will print.  Any number of separators can be chained!  This
allows for glitter to avoid annoying sequences of literals in many
cases, and makes it much more natural to space out different parts
of the prompt.

Separators will appear if _any_ non-separator expression in the
tree was written before, but separators only apper _between_ the
expected expressions.  For example:

    "b@B 'test'"

If b _and_ B print, then the @ will print.  If only b prints, the
space will still appear between b and test.  This should be the
most intuitive behaviour.