Add error messages to parser
Getting the error messages I wanted required using custom
combinators; however, that was not much of a problem as nom is
shockingly easy to extend and use!

Error messages required changing some things around, but most of
the system relies on cleverly placed context combinators and
liberal use of map_err and map_fail to create more specific errors
when necessary.