Experiment with delimited version of take_until
Works well except that it becomes over-complicated when dealing
with closing or opening braces inside of strings.  Would need to
be aware of _all_ delimited pairs and create some sort of token
tree ala rust in order to function properly.  Has many benefits
in terms of finding all matching braces, etc, and is worth
re-visiting if I experiment with custom input types in nom.

The input type could potentially be defined as:

```
enum<'a> TokenTree {
    Quote(Vec<TokenTree<'a>>),
    Parens(Vec<TokenTree<'a>>),
    Brace(Vec<TokenTree<'a>>),
    Curly(Vec<TokenTree<'a>>),
    Angle(Vec<TokenTree<'a>>),
    Leaf(&'a str),
}
```

This would add a tokenize stage which builds a Vec of Token Trees
from the input.